+++ packages/js/src/types.ts
@@ -1,3 +1,5 @@
+export type AsyncTracker = any
+
 export interface CustomDimension {
   id: number
   value: string
@@ -75,3 +77,8 @@ export interface SetEcommerceViewParams {
   productCategory?: string
   productPrice?: number
 }
+
+export type EcommerceItems = Record<
+  number,
+  [string, string, string, number, number]
+>